# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of water app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://the-dominators-back-project.onrender.com"
    }
  ],
...
  "paths": {
...
    "/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User logout from the app",
        "operationId": "logOutUser",
        "description": "Logs out a user from the app by clearing the session and authentication tokens",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {}
        },
        "responses": {
          "204": {
            "description": "Successfully logged out the user"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Session not found"
                    },
                    "data": {
                      "type": "string",
                      "example": "Session not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}